+++ src/main/java/dansplugins/democracy/commands/RunCommand.java
@@ -1,7 +1,13 @@
 package dansplugins.democracy.commands;
 
+import dansplugins.democracy.Democracy;
+import dansplugins.democracy.data.PersistentData;
+import dansplugins.democracy.factories.CandidateFactory;
+import dansplugins.democracy.objects.Election;
+import dansplugins.factionsystem.externalapi.MF_Faction;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 import preponderous.ponder.minecraft.bukkit.abs.AbstractPluginCommand;
 
 import java.util.ArrayList;
@@ -12,22 +18,49 @@ import java.util.Arrays;
  * @author Daniel McCoy Stephenson
  */
 public class RunCommand extends AbstractPluginCommand {
+    private final Democracy democracy;
+    private final PersistentData persistentData;
+    private final CandidateFactory candidateFactory;
 
-    public RunCommand() {
+    public RunCommand(Democracy democracy, PersistentData persistentData, CandidateFactory candidateFactory) {
         super(new ArrayList<>(Arrays.asList("run")), new ArrayList<>(Arrays.asList("d.run")));
+        this.democracy = democracy;
+        this.persistentData = persistentData;
+        this.candidateFactory = candidateFactory;
     }
 
     @Override
     public boolean execute(CommandSender commandSender) {
-        // TODO: implement
-        commandSender.sendMessage(ChatColor.RED + "This command isn't implemented yet.");
-        return false;
+        if (!(commandSender instanceof Player)) {
+            commandSender.sendMessage("This command cannot be used in the console.");
+            return false;
+        }
+        Player player = (Player) commandSender;
+
+        MF_Faction faction = democracy.getMedievalFactionsAPI().getFaction(player);
+        if (faction == null) {
+            player.sendMessage(ChatColor.RED + "You must be in a faction to run in an election.");
+            return false;
+        }
+
+        Election election = persistentData.getElectionForFaction(faction.getName());
+        if (election == null) {
+            player.sendMessage(ChatColor.RED + "There is no election currently in progress in your faction.");
+            return false;
+        }
+
+        if (election.isCandidate(player.getUniqueId())) {
+            player.sendMessage(ChatColor.RED + "You are already a candidate in this election.");
+            return false;
+        }
+
+        candidateFactory.createCandidate(player, election);
+        player.sendMessage(ChatColor.GREEN + "You are now a candidate in the election.");
+        return true;
     }
 
     @Override
     public boolean execute(CommandSender commandSender, String[] strings) {
-        // TODO: implement
-        commandSender.sendMessage(ChatColor.RED + "This command isn't implemented yet.");
-        return false;
+        return execute(commandSender);
     }
 }